Fix bug in --ask option of CLI. Fixes #62
When we reversed the order of the filenames in the .approvals file,
we forgot to update the CLI, which relied on the order of arguments for
the mv command.

Instead of overwriting the approval file with the newly received one,
it did the reverse, which caused the approval file to disappear
altogether.