Avoid redefining the write method
The `attr_accessor :write` line was defining a getter method named
`write` which was then being overridden by the `write` method that is
part of the `Writer` protocol.

It seems like there was no need for any of the variables to have
accessors, so I removed them.  It is possible that I don't fully
understand the purpose of those accessors, though, so this commit may
not be the right way to solve the problem.