Fix bug preventing array and hash values from being filtered
The filtering of keys would not work correctly if the corresponding values for the keys were Arrays or Hashes. This was caused by `censored` method checking for the type of the value before checking if the key matched any of the filters. By flipping around the logic and checking if the key should be filtered before checking the value's type, filtering works correctly.

Test plan: I added two new filter_spec tests to ensure this behavior is working correctly.