Install a more recent version of bundler on CI
On some Ruby versions, an older version of bundler is installed.  Some of these older versions have a [bug](https://github.com/bundler/bundler/issues/3558) that causes our builds to fail.  Of recent note are some failures on Ruby 1.9.3 with bundler 1.7.6.

According to the [related issue on TravisCI](https://github.com/travis-ci/travis-ci/issues/3531), the solution is to add a `before_install` step to the build to update bundler.