Use latest Rubinius (#80)
There don’t seem to binary versions of rbx-2 available on Travis any
more, so switch to using `rbx` instead.  Per https://github.com/travis-ci/docs-travis-ci-com/pull/740, this seems to be the correct thing to do.

However, the `rbx` build is still failing.  This seems like it might be a temporary build environment issue, so allow the rbx build to fail until it is resolved.